move some classes to another packages
