+++ src/main/kotlin/data/datasource/weather/WeatherDataSource.kt
@@ -1,4 +1,4 @@
-package org.example.data.datasource
+package org.example.data.datasource.weather
 
 import WeatherDto
 import logic.model.LocationModel
