+++ src/main/kotlin/data/datasource/weather/WeatherDataSourceImpl.kt
@@ -1,4 +1,4 @@
-package org.example.data.datasource
+package org.example.data.datasource.weather
 
 import WeatherDto
 import io.ktor.client.*
