+++ src/main/kotlin/data/repository/WeatherRepositoryImpl.kt
@@ -2,7 +2,7 @@ package org.example.data.repository
 
 import logic.model.LocationModel
 import logic.model.WeatherModel
-import org.example.data.datasource.WeatherDataSource
+import org.example.data.datasource.weather.WeatherDataSource
 import org.example.data.mapper.toWeatherModelList
 import org.example.logic.exceptions.NoWeatherFoundException
 import org.example.logic.repository.WeatherRepository
